Max latency computation

+++ CAESAR/src/event/Event.java
@@ -1,7 +1,7 @@
 package event;
 
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * An event has a type, time stamp and carries a vehicle identifier. 
@@ -29,7 +29,7 @@ public abstract class Event {
 	 * @param s				type of complex event
 	 * @param distrProgr 	distributor progress in application time of input event
 	 */
-	public void printError (PositionReport p, double emit, AtomicBoolean failed, AtomicLong max_latency, String s, long distrProgr) {
+	public void printError (PositionReport p, double emit, AtomicBoolean failed, AtomicInteger max_latency, String s, long distrProgr) {
 		
 		// Print an error message and update the accident warning failed variable
 		int diff = new Double(emit).intValue() - new Double(p.sec).intValue();
