+++ CAESAR/src/iogenerator/Main.java
@@ -6,8 +6,6 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
-
 import run.*;
  
 public class Main {
@@ -77,7 +75,7 @@ public class Main {
 		ArrayList<CountDownLatch> dones = new ArrayList<CountDownLatch>();
 		ArrayList<HashMap<RunID,Run>> runtables = new ArrayList<HashMap<RunID,Run>>();
 		
-		AtomicLong max_latency = new AtomicLong(0);
+		AtomicInteger max_latency = new AtomicInteger(0);
 		
 		for (int i=12; i<args.length; i++) {
