+++ CAESAR/src/scheduler/RunAndQueryDrivenScheduler.java
@@ -5,7 +5,6 @@ import java.util.HashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
 import run.*;
 import distributor.*;
 import iogenerator.*;
@@ -24,7 +23,7 @@ public class RunAndQueryDrivenScheduler extends Scheduler implements Runnable {
 	int LP_frequency;
 	
 	public RunAndQueryDrivenScheduler (AtomicInteger dp, HashMap<Double,Long> distrProgrPerSec, HashMap<RunID,Run> rs, RunQueues rq, RunQueues hprq,ExecutorService e, 
-								CountDownLatch tn, CountDownLatch d, ArrayList<XwayDirPair> xds, int lastS, long start, AtomicLong max_late, AtomicInteger x0, AtomicInteger x1, int hpf, int lpf) {		
+								CountDownLatch tn, CountDownLatch d, ArrayList<XwayDirPair> xds, int lastS, long start, AtomicInteger max_late, AtomicInteger x0, AtomicInteger x1, int hpf, int lpf) {		
 		super(dp,distrProgrPerSec,rs,rq,e,tn,d,xds,lastS,start,max_late);
 		HPrunqueues = hprq;
 		xway0dir0HPrunsFromSeg = x0;
